Base layer of stoppable behavior
